[diffusion, tests] refactor: move compile helpers to utils
Keep regional compilation policy and its Diffusers eager-boundary workaround out of the FSDP engine implementation. Isolate wiring tests from the process-global attention helper patch while retaining explicit call verification.
